Word_Count_I_Refactor_And_Clean_Code
Instead for for-loop we are now returning the number of words with only letters in sentence with Java streams.
Also solving couple of sonar warnings and delete unused classes and fields in Sentence class.

+++ src/main/java/SentenceWordsCounting.java
@@ -1,4 +1,3 @@
 public interface SentenceWordsCounting {
-
-    int countingWordsOnlyLetter(String sentence);
+    long countingWordsOnlyLetter(String sentence);
 }
